On branch Luis  Changes to be committed: 	modified:   multiply.h

+++ multiply.h
@@ -1,7 +1,6 @@
 #ifndef MULTIPLY_H
 #define MULTIPLY_H
-
-    double multi(double num1, double num2) {
+    int multi(int num1, int num2) {
         return num1 * num2;
     };
 #endif
